Create AAE task2 code.py
Task2 is designed based on the code of task1. I chose the area of y of range (25, 30) as the minus-cost-area. According to the design of task1, this area can cover the most paths, so that the total trip time required is the least. First define this area, then let "self.Delta_C3"=-0.05, and finally define the range of this area. Finally, the computer gave a conclusion: Total trip time required=75.676548008344